Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-    int a = 3.14;
-    scanf("%f", &a);
+    float a = 3.14;
+    scanf("%.2f", &a);
     printf(" You entered: %f", a);
     return 0;
 }
